tests/node_metrics: factor out metrics helpers into utils class
